+++ languages/ca.js
@@ -1,0 +1,61 @@
+/*!
+ * numbro.js language configuration
+ * language : Catalan
+ * author : David Gölzhäuser : https://github.com/idoodler
+ */
+
+module.exports = {
+    languageTag: "ca",
+    delimiters: {
+        thousands: ".",
+        decimal: ","
+    },
+    abbreviations: {
+        thousand: "k",
+        million: "mm",
+        billion: "b",
+        trillion: "t"
+    },
+    ordinal: (number) => {
+        if (number >= 5) {
+            return "è";
+        }
+        let b = number % 10;
+        return (b === 1 | b === 3) ? "r" : (b === 2) ? "n" : (b === 4) ? "t" : "è";
+    },
+    currency: {
+        symbol: "€",
+        position: "postfix",
+        code: "EUR"
+    },
+    currencyFormat: {
+        thousandSeparated: true,
+        totalLength: 4,
+        spaceSeparated: true,
+        spaceSeparatedCurrency: true,
+        average: true
+    },
+    formats: {
+        fourDigits: {
+            totalLength: 4,
+            spaceSeparated: true,
+            average: true
+        },
+        fullWithTwoDecimals: {
+            output: "currency",
+            mantissa: 2,
+            spaceSeparated: true,
+            thousandSeparated: true
+        },
+        fullWithTwoDecimalsNoCurrency: {
+            mantissa: 2,
+            thousandSeparated: true
+        },
+        fullWithNoDecimals: {
+            output: "currency",
+            spaceSeparated: true,
+            thousandSeparated: true,
+            mantissa: 0
+        }
+    }
+};